Move files to typescript
