+++ src/RNGalleryList.ts
@@ -1,6 +1,5 @@
 'use strict';
 import React from 'react';
-import { View } from 'react-native';
 const WelcomePage = require('./WelcomePage');
 const CheckBoxExamplePage = require('./examples/CheckBoxExamplePage');
 
@@ -8,7 +7,7 @@ interface IRNGalleryExample {
   key: string;
   component: React.Component;
 }
-  
+
 export const RNGalleryList: Array<IRNGalleryExample> = [
   {
     key: 'Wecome',
