+++ src/examples/CheckBoxExamplePage.tsx
@@ -1,0 +1,22 @@
+'use strict';
+import {StyleSheet, View, Text} from 'react-native';
+import React from 'react';
+import CheckBox from '@react-native-community/checkbox';
+
+class CheckBoxExamplePage extends React.Component {
+  render() {
+    return (
+      <View
+        style={{
+          height: '100%',
+          justifyContent: 'center',
+          alignItems: 'center',
+        }}>
+        <Text>Sample Checkbox</Text>
+        <CheckBox />
+      </View>
+    );
+  }
+}
+
+export default CheckBoxExamplePage;
